3 December 2015
An example .csv log file has been uploaded with all the values recorded
during the locking routine.

A basic and slightly more advanced log viewer files have been included
as well to view the files.  Although it is advised for any user to view
the files as best as they see fit.